requirements.txt
tensorflow==2.9.1
tensorflow_datasets==4.6.0
PyYAML>=6.0
wandb>=0.12.17
matplotlib>=3.5.2